Create Main.java
The code, import the sanner library, then uses if-else statements to check whether the number is positive, negative, or zero, and prints the corresponding message.

After that, a for loop is used to print the numbers from 1 to the entered number. The loop iterates from 1 to number, incrementing the i variable by 1 in each iteration, and prints the value of i.